started to convert magnitude representation to int16 integer values

+++ lib/smaudio.hh
@@ -27,7 +27,7 @@ class AudioBlock
 public:
   std::vector<float> noise;          //!< noise envelope, representing the original signal minus sine components
   std::vector<float> freqs;          //!< frequencies of the sine components of this frame
-  std::vector<float> mags;           //!< magnitudes of the sine components
+  std::vector<int16_t> mags;         //!< magnitudes of the sine components
   std::vector<float> phases;         //!< phases of the sine components
   std::vector<float> lpc_lsf_p;      //!< LPC line spectrum frequencies, P(z) roots
   std::vector<float> lpc_lsf_q;      //!< LPC line spectrum frequencies, Q(z) roots
